Extract DHM parsing to own file
